Added SV_AddResource_hook, fixed chain EV_Precache

+++ rehlds/public/rehlds/rehlds_api.h
@@ -244,8 +244,8 @@ typedef IHookChain<int, const char *> IRehldsHook_PF_precache_sound_I;
 typedef IHookChainRegistry<int, const char *> IRehldsHookRegistry_PF_precache_sound_I;
 
 //EV_Precache hook
-typedef IHookChain<int, int, const char *> IRehldsHook_EV_Precache;
-typedef IHookChainRegistry<int, int, const char *> IRehldsHookRegistry_EV_Precache;
+typedef IHookChain<unsigned short, int, const char *> IRehldsHook_EV_Precache;
+typedef IHookChainRegistry<unsigned short, int, const char *> IRehldsHookRegistry_EV_Precache;
 
 //SV_AddResource hook
 typedef IVoidHookChain<void, resourcetype_t, const char *, int, unsigned char, int> IRehldsHook_SV_AddResource;
